Change "motor" peripheral name to "pwm", change brightness data type to Int.

+++ app/src/main/java/com/maverick/iotsocket/model/Sensor.kt
@@ -1,3 +1,3 @@
 package com.maverick.iotsocket.model
 
-data class Sensor(val temperature: Float, val pressure: Float, val brightness: Float)
+data class Sensor(val temperature: Float, val pressure: Float, val brightness: Int)
